requirements.txt
psycopg2-binary==2.9.10
python-dotenv==1.0.1
termcolor==2.5.0
colorama==0.4.6
django-apscheduler==0.7.0
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.0
django-extensions==3.2.3
Django~=5.1.5
gunicorn==23.0.0
requests==2.31.0
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.10.1
